my_nunny/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Mail
Flask-WTF
itsdangerous
PyJWT # For token generation in models
Flask-Bootstrap4
SQLAlchemy<2.0 # Pin for compatibility with Flask-Bootstrap4 examples / older Flask-SQLAlchemy
python-dotenv
